Create _utils submodule to store different utilities used across diferent functions.
