+++ wiutils/_dwc/__init__.py
@@ -1,1 +1,1 @@
-from wiutils._dwc import constants, mapping, optional, order, utils
+from wiutils._dwc import constants, mapping, optional, order
