--- wiutils/_dwc/nls/__init__.py
@@ -1,1 +1,0 @@
-from wiutils._dwc.nls import es
