+++ wiutils/_nls/__init__.py
@@ -1,0 +1,1 @@
+from wiutils._nls import es
