+++ wiutils/_utils/__init__.py
@@ -1,0 +1,1 @@
+from wiutils._utils import data, language, taxonomy
